backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
motor==3.3.2
pymongo==4.6.3
pydantic==2.6.4
pandas==2.2.1
python-multipart==0.0.9
python-dotenv==1.0.1
openai>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
fpdf2>=2.7.8
email-validator>=2.0.0
